feat:add migrations, add headers

+++ src/jobs/jobs.controller.ts
@@ -8,6 +8,7 @@ import {
   Body,
   UsePipes,
   Headers,
+  UnauthorizedException,
 } from '@nestjs/common';
 
 import {
@@ -29,63 +30,108 @@ import {
 import toArray from '../commons/toArray';
 import { ZodValidationPipe } from 'nestjs-zod';
 import { JobsService } from './jobs.service';
+import { Authenticator } from 'src/commons/tokenGenerator';
 
 @Controller('/job')
 export class JobsController {
-  constructor(private readonly appService: JobsService) {}
+  constructor(
+    private readonly appService: JobsService,
+    private readonly auth: Authenticator,
+  ) {}
 
-  @Get('/health')
+  @Get('')
   getHello(): string {
     return this.appService.getHello();
   }
   @Get('/offer/:id')
   @UsePipes(new ZodValidationPipe(idParam))
-  async getOneJob(@Param('id') id: JobId) {
+  async getOneJob(
+    @Param('id') id: JobId,
+    @Headers('Authorization') authorization: string,
+  ) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.getOneJob(id);
   }
   @Get('/offers')
   async getJobs(
     @Body() body: JobIds[],
-    @Headers('Authorization') Authorization: string,
+    @Headers('Authorization') authorization: string,
   ) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.getJobs(toArray(body));
   }
   @Get('/international')
   @UsePipes(new ZodValidationPipe(isBoolean))
-  async getNationalOrInternationalJobs(@Body('international') body: IsBoolean) {
+  async getNationalOrInternationalJobs(
+    @Body('international') body: IsBoolean,
+    @Headers('Authorization') authorization: string,
+  ) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.getNationalAndInternationalJobs(body);
   }
   @Get('/cities')
   @UsePipes(new ZodValidationPipe(isString))
-  async getCities(@Body('country') body: IsString) {
+  async getCities(
+    @Body('country') body: IsString,
+    @Headers('Authorization') authorization: string,
+  ) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.getCities(body);
   }
   @Get('/countries')
-  async getCountries() {
+  async getCountries(@Headers('Authorization') authorization: string) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.getCountries();
   }
   @Get('/total')
-  async getSumJobsPerCountry() {
+  async getSumJobsPerCountry(@Headers('Authorization') authorization: string) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.getTotalJobsPerCountry();
   }
   @Post('/create')
   @UsePipes(new ZodValidationPipe(jobBodyArray))
-  async createJobs(@Body() body: JobBody[]): Promise<void> {
+  async createJobs(
+    @Body() body: JobBody[],
+    @Headers('Authorization') authorization: string,
+  ): Promise<void> {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.createJobs(toArray(body));
   }
   @Put('/update')
   @UsePipes(new ZodValidationPipe(jobUpdateBodyArray))
-  async updateJobs(@Body() body: JobUpdateBody): Promise<void> {
+  async updateJobs(
+    @Body() body: JobUpdateBody,
+    @Headers('Authorization') authorization: string,
+  ): Promise<void> {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.updateJobs(body);
   }
   @Post('/filter')
   @UsePipes(new ZodValidationPipe(bodyFilter))
-  async filterJobs(@Body() body: FilterBody) {
+  async filterJobs(
+    @Body() body: FilterBody,
+    @Headers('Authorization') authorization: string,
+  ) {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return this.appService.filterJobs(body);
   }
   @Delete('/delete')
   @UsePipes(new ZodValidationPipe(jobIds))
-  async deleteJobs(@Body() body: JobIds): Promise<void> {
+  async deleteJobs(
+    @Body() body: JobIds,
+    @Headers('Authorization') authorization: string,
+  ): Promise<void> {
+    if (!this.auth.verifyTokenData(authorization))
+      throw new UnauthorizedException();
     return await this.appService.deleteJobs(toArray(body));
   }
 }
